refactor: restructure IAM module to utilize TokenModule for JWT management
- Replaced direct JWT handling in AuthenticationService with TokenService from the new TokenModule.
- Updated IamModule to import TokenModule, enhancing modularity and separation of concerns.
- Refactored AccessTokenGuard to use TokenService for token verification, simplifying its implementation.
- Added TokenService and its tests to manage JWT generation and verification more effectively.
