+++ src/iam/token/token.module.ts
@@ -1,0 +1,15 @@
+import { Module } from '@nestjs/common';
+import { JwtModule } from '@nestjs/jwt';
+import { TokenService } from './token.service';
+import { ConfigModule } from '@nestjs/config';
+import jwtConfig from '../config/jwt.config';
+
+@Module({
+  imports: [
+    JwtModule.registerAsync(jwtConfig.asProvider()),
+    ConfigModule.forFeature(jwtConfig),
+  ],
+  providers: [TokenService],
+  exports: [TokenService],
+})
+export class TokenModule {}
